firstVersion
SearchPage and Api Routes

+++ app/lib/models/movie_data.dart
@@ -76,9 +76,9 @@ class Movies {
     releaseDate = json['release_date'];
     title = json['title'];
     video = json['video'];
-    voteAverage = json['vote_average'];
+    voteAverage = json['vote_average'].toDouble();
     voteCount = json['vote_count'];
-    popularity = json['popularity'];
+    popularity = json['popularity'].toDouble();
     mediaType = json['media_type'];
   }
